Override secret with env variable (if present).
To avoid hard-coding the application secret (which is used for
encrypting sessions) in the configuration file, it can be supplied
via an environment variable. If the env variable is present it will
be used, otherwise it will fall back on the default secret.

More info: https://www.playframework.com/documentation/2.3.x/ApplicationSecret